reference components in other components using BRef

+++ Boon/include/Networking/Components/NetRigidbody2D.h
@@ -76,7 +76,7 @@ namespace Boon
 			else
 			{
 				Rigidbody2D& rb = gameObject.GetComponent<Rigidbody2D>();
-				rb.Type = Rigidbody2D::BodyType::Kinematic;
+				rb.Type = (int)Rigidbody2D::BodyType::Kinematic;
 				glm::vec3 pos = { DequantizePos(QPosX), DequantizePos(QPosY), DequantizePos(QPosZ) };
 				float rot = DequantizeAngleDeg(QRotDeg);
 				glm::vec2 vel = { DequantizePos(QVelX), DequantizePos(QVelY) };
